+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Metrics perf improvements, bug fixes.
+  Replace MetricProcessor with MetricReader.
+  ([#2306](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2306))
+
 * Add `BatchExportActivityProcessorOptions` which supports field value overriding
   using `OTEL_BSP_SCHEDULE_DELAY`, `OTEL_BSP_EXPORT_TIMEOUT`,
   `OTEL_BSP_MAX_QUEUE_SIZE`, `OTEL_BSP_MAX_EXPORT_BATCH_SIZE`